Exposes public types on root namespace.
Improves ease of usage.
