+++ samples/NScatterGather.Samples/Samples/1.HelloWorld.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Threading.Tasks;
-using NScatterGather.Recipients;
 
 namespace NScatterGather.Samples
 {
