+++ samples/NScatterGather.Samples/Samples/2.FilterOnResponse.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Threading.Tasks;
-using NScatterGather.Recipients;
 
 namespace NScatterGather.Samples.Samples
 {
