+++ samples/NScatterGather.Samples/Samples/3.InvokeAsyncMethods.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Threading.Tasks;
-using NScatterGather.Recipients;
 
 namespace NScatterGather.Samples.Samples
 {
