+++ samples/NScatterGather.Samples/Samples/5.Timeout.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using NScatterGather.Recipients;
 
 namespace NScatterGather.Samples.Samples
 {
