+++ src/NScatterGather/Invocations/CompletedInvocation.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 
-namespace NScatterGather.Invocations
+namespace NScatterGather
 {
     public class CompletedInvocation<TResponse>
     {
