+++ src/NScatterGather/Invocations/FaultedInvocation.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace NScatterGather.Invocations
+namespace NScatterGather
 {
     public class FaultedInvocation
     {
