+++ src/NScatterGather/Responses/AggregatedResponseExtensions.cs
