+++ tests/NScatterGather.Tests/Inspection/TypeInspectorTests.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Xunit;
+﻿using Xunit;
 
 namespace NScatterGather.Inspection
 {
